+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `unwrap` operator.
 
+* Fix `Summarize.value_by` when using `symbolize: true` and a default value.
+
 # 0.18.5 - 2021-06-08
 
 * Add a Summarize.value_by that allows flipping vertical series to a